refactoring to separate concerns for exec and static files
